Fixed Stack - using Array based problem
Write a program to perform the following operations using Java.
1. Read the count of elements from user to create the stack using array. 
2. Create the class and functions for the following and call them using a menu based program.
a. Display the stack
b. Push the data on the stack
c. Pop the data from the stack
d. Check how many elements are there on the stack.
e. Check to see if the stack is full?
f. Check to see if stack is empt